Added function to check for boundary closure.
Added function to check for boundary closure to know when to stop the mapping, and to record the elapsed time into a file named maptime.txt.

Since the global variable occdata was not being used by other functions, we have co-opted it to store the 2D map data that we will pass to the closure function. As a result, code was also added to convert the 1D map data to 2D, and to make all the values positive.